requirements.txt
packaging==24.1
pillow==10.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pytesseract==0.3.10
python-decouple==3.8
pyzbar==0.1.9
